+++ rich/console.py
@@ -984,14 +984,18 @@ class Console:
                 else:
                     text = self._render_buffer()
                     if text:
-                        if WINDOWS:  # pragma: no cover
-                            # https://bugs.python.org/issue37871
-                            write = self.file.write
-                            for line in text.splitlines(True):
-                                write(line)
-                        else:
-                            self.file.write(text)
-                        self.file.flush()
+                        try:
+                            if WINDOWS:  # pragma: no cover
+                                # https://bugs.python.org/issue37871
+                                write = self.file.write
+                                for line in text.splitlines(True):
+                                    write(line)
+                            else:
+                                self.file.write(text)
+                            self.file.flush()
+                        except UnicodeEncodeError as error:
+                            error.reason = f"{error.reason}\n*** You may need to add PYTHONIOENCODING=utf-8 to your environment ***"
+                            raise
 
     def _render_buffer(self) -> str:
         """Render buffered output, and clear buffer."""
@@ -1229,7 +1233,7 @@ if __name__ == "__main__":  # pragma: no cover
 
     console.log("Hello, World!", "{'a': 1}", repr(console))
 
-    console.log(
+    console.print(
         {
             "name": None,
             "empty": [],