+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [5.0.1] - Unreleased
+## [5.1.0] - 2020-08-08
+
+### Added
+
+- Added Text.cell_len
+- Added helpful message regarding unicode decoding errors
 
 ### Fixed
 
 - Fixed deprecation warnings re backslash https://github.com/willmcgugan/rich/issues/210
 - Fixed repr highlighting of scientific notation, e.g. 1e100
+
+### Changed
+
+- Implemented pretty printing, and removed pprintpp from dependancies
+- Optimized Text.join
 
 ## [5.0.0] - 2020-08-02